services/privacy-shield/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-cloud-vision==3.7.1
Pillow==10.1.0
numpy==1.24.3
